requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
Markdown==3.3.7
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
whitenoise==6.2.0
